+++ src/main/java/com/naocraftlab/foggypalegarden/integration/ModMenuIntegration.java
@@ -1,0 +1,26 @@
+package com.naocraftlab.foggypalegarden.integration;
+
+import com.naocraftlab.foggypalegarden.gui.ClothConfigScreen;
+import com.naocraftlab.foggypalegarden.gui.NoClothConfigScreen;
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.client.gui.screen.Screen;
+
+public final class ModMenuIntegration implements ModMenuApi {
+
+    @Override
+    public ConfigScreenFactory<Screen> getModConfigScreenFactory() {
+        return parent -> {
+            if (isClothConfigPresent()) {
+                return ClothConfigScreen.of(parent);
+            } else {
+                return NoClothConfigScreen.of(parent);
+            }
+        };
+    }
+
+    private static boolean isClothConfigPresent() {
+        return FabricLoader.getInstance().isModLoaded("cloth-config");
+    }
+}